// src/core/enums/error-names.ts
export enum ErrorNames {
    TITLE_MISSING_ERROR = "Title is required",
    TITLE_TYPE_ERROR = "Title must be string",
    TITLE_LENGTH_ERROR = "Title must be between 2 and 40 symbols",
    AUTHOR_MISSING_ERROR = "Author is required",
    AUTHOR_TYPE_ERROR = "Author must be a string",
    AUTHOR_LENGTH_ERROR = "Author must be between 2 and 20 symbols",
    AVAILABLE_RESOLUTIONS_MISSING_ERROR = "availableResolutions is required",
    AVAILABLE_RESOLUTIONS_TYPE_ERROR = "availableResolutions must be array",
    AVAILABLE_RESOLUTIONS_LENGTH_ERROR = "Invalid availableResolutions length",
    AVAILABLE_RESOLUTIONS_VALUE_ERROR = "This resolution is not supported: ",
    CAN_BE_DOWNLOADED_MISSING_ERROR = "canBeDownloaded is required",
    CAN_BE_DOWNLOADED_TYPE_ERROR = "canBeDownloaded must be boolean",
    MIN_AGE_RESTRICTION_MISSING_ERROR = "minAgeRestriction is required",
    MIN_AGE_RESTRICTION_TYPE_ERROR = "minAgeRestriction must be number or null",
    MIN_AGE_RESTRICTION_RANGE_ERROR = "minAgeRestriction must be between 1 and 18",
    PUBLICATION_DATE_MISSING_ERROR = "publicationDate is required",
    PUBLICATION_DATE_TYPE_ERROR = "publicationDate must be a string",
    PUBLICATION_DATE_FORMAT_ERROR = "publicationDate must be in ISO format",
}